Removed shared folder from tools. Everything is now in Utils. Tool specific shared utils are packaged and namespaced. This can also be imported by plugins that wish to build a modified tool.
