+++ src/tools/EllipticalRoiTool.js
@@ -14,7 +14,7 @@ import pointInEllipse from '../util/pointInEllipse.js';
 import calculateEllipseStatistics from '../util/calculateEllipseStatistics.js';
 import calculateSUV from '../util/calculateSUV.js';
 //
-import numberWithCommas from './shared/numbersWithCommas.js';
+import numbersWithCommas from '../util/numbersWithCommas.js';
 
 /**
  * @export @public @class
@@ -334,11 +334,11 @@ export default class EllipticalRoiTool extends BaseAnnotationTool {
         }
 
         // Create a line of text to display the mean and any units that were specified (i.e. HU)
-        let meanText = `Mean: ${numberWithCommas(
+        let meanText = `Mean: ${numbersWithCommas(
           meanStdDev.mean.toFixed(2)
         )}${moSuffix}`;
         // Create a line of text to display the standard deviation and any units that were specified (i.e. HU)
-        let stdDevText = `StdDev: ${numberWithCommas(
+        let stdDevText = `StdDev: ${numbersWithCommas(
           meanStdDev.stdDev.toFixed(2)
         )}${moSuffix}`;
 
@@ -346,9 +346,9 @@ export default class EllipticalRoiTool extends BaseAnnotationTool {
         if (meanStdDevSUV && meanStdDevSUV.mean !== undefined) {
           const SUVtext = ' SUV: ';
 
-          meanText += SUVtext + numberWithCommas(meanStdDevSUV.mean.toFixed(2));
+          meanText += SUVtext + numbersWithCommas(meanStdDevSUV.mean.toFixed(2));
           stdDevText +=
-            SUVtext + numberWithCommas(meanStdDevSUV.stdDev.toFixed(2));
+            SUVtext + numbersWithCommas(meanStdDevSUV.stdDev.toFixed(2));
         }
 
         // Add these text lines to the array to be displayed in the textbox
@@ -368,7 +368,7 @@ export default class EllipticalRoiTool extends BaseAnnotationTool {
         }
 
         // Create a line of text to display the area and its units
-        const areaText = `Area: ${numberWithCommas(area.toFixed(2))}${suffix}`;
+        const areaText = `Area: ${numbersWithCommas(area.toFixed(2))}${suffix}`;
 
         // Add this text line to the array to be displayed in the textbox
         textLines.push(areaText);
