+++ src/tools/RotateTool.js
@@ -1,7 +1,7 @@
 /* eslint no-loop-func: 0 */ // --> OFF
 import external from './../externalModules.js';
 import BaseTool from './../base/BaseTool.js';
-import angleBetweenPoints from './shared/angleBetweenPoints.js';
+import angleBetweenPoints from '../util/angleBetweenPoints.js';
 
 /**
  * @export @public @class
