+++ src/tools/ZoomTool.js
@@ -1,7 +1,7 @@
 import external from '../externalModules.js';
 import BaseTool from '../base/BaseTool.js';
 import { clipToBox } from '../util/clip.js';
-import { correctShift, changeViewportScale } from './shared/zoom.js';
+import { correctShift, changeViewportScale } from '../util/zoom.js';
 
 /**
  * @export @public @class
