+++ src/util/brush/index.js
@@ -1,0 +1,8 @@
+import { drawBrushPixels, drawBrushOnCanvas } from './drawBrush.js';
+import getCircle from './getCircle.js';
+
+export default {
+  drawBrushPixels,
+  drawBrushOnCanvas,
+  getCircle
+};
