+++ src/util/freehand/FreehandLineFinder.js
@@ -1,15 +1,19 @@
-import { getToolState } from '../../../stateManagement/toolState.js';
-import external from '../../../externalModules.js';
+import { getToolState } from '../../stateManagement/toolState.js';
+import external from '../../externalModules.js';
 import { ClickedLineData } from './ClickedLineData.js';
 
 const toolType = 'freehandMouse';
 const distanceThreshold = 10;
 
-export class FreehandLineFinder {
-  /* eslint no-underscore-dangle: ["error", { "allowAfterThis": true }] */
-
+/**
+ * @export @public @class
+ * @name FreehandLineFinder
+ * @classdesc Class that finds lines of freehand ROIs based on click proximity.
+ */
+export default class FreehandLineFinder {
   /**
   * Constructs a linefinder with the eventdata
+  * @public @method
   *
   * @param {Object} eventData - Data object associated with the event.
   */
@@ -19,9 +23,9 @@ export class FreehandLineFinder {
 
   /**
   * Looks for lines near the mouse cursor.
+  * @public @method
   *
-  *  @public
-  *  @return {ClickedLineData}
+  * @returns {ClickedLineData} Data object corresponding to the clicked line.
   */
   findLine () {
     const closestToolIndex = this.findTool();
@@ -45,9 +49,9 @@ export class FreehandLineFinder {
 
   /**
   * Looks for tools near the mouse cursor.
+  * @public @method
   *
-  *  @public
-  *  @return {ClickedLineData}
+  * @returns {ClickedLineData} Data object corresponding to the clicked line.
   */
   findTool () {
     this._toolData = getToolState(this._eventData.element, toolType);
@@ -64,8 +68,9 @@ export class FreehandLineFinder {
 
   /**
   * Finds the nearest handle to the mouse cursor for all tools.
+  * @private @method
   *
-  * @return {Object} closestHandle - The handle closest to the point.
+  * @returns {object} The handle closest to the point.
   */
   _nearestHandleToPointAllTools () {
     const toolData = this._toolData;
@@ -93,9 +98,10 @@ export class FreehandLineFinder {
 
   /**
   * Finds the nearest handle to the mouse cursor for a specific tool.
+  * @private @method
   *
-  * @param {Number} toolIndex - The index of the particular freehand tool.
-  * @return {Object} An object containing information about the closest handle.
+  * @param {number} toolIndex The index of the particular freehand tool.
+  * @return {object} An object containing information about the closest handle.
   */
   _nearestHandleToPoint (toolIndex) {
     const eventData = this._eventData;
@@ -132,9 +138,10 @@ export class FreehandLineFinder {
 
   /**
   * Finds all the lines close to the mouse point for a particular tool.
+  * @private @method
   *
-  * @param {Number} toolIndex - The index of the particular freehand tool.
-  * @return {Object} An array of lines close to the mouse point.
+  * @param {number} toolIndex The index of the particular freehand tool.
+  * @returns {object} An array of lines close to the mouse point.
   */
   _getCloseLinesInTool (toolIndex) {
     const toolData = this._toolData;
@@ -158,11 +165,12 @@ export class FreehandLineFinder {
   }
 
   /**
-  * Finds the line the user clicked on from an array of close lines.
+  * Finds the line the user clicked on from an array of close lines.\
+  * @private @method
   *
-  * @param {Number} toolIndex - The index of the particular freehand tool.
-  * @param {Object} closeLines - An array of lines close to the mouse point.
-  * @return {ClickedLineData|null} An instance of ClickedLineData containing information about the line, or null if no line is correct.
+  * @param {number} toolIndex The index of the particular freehand tool.
+  * @param {object} closeLines An array of lines close to the mouse point.
+  * @returns {ClickedLineData|null} An instance of ClickedLineData containing information about the line, or null if no line is correct.
   */
   _findCorrectLine (toolIndex, closeLines) {
     // Test if any candidate lines can be projected onto by the mousePoint
@@ -178,10 +186,11 @@ export class FreehandLineFinder {
 
   /**
   * Returns true if the mouse point projects onto the line segment.
+  * @private @method
   *
-  * @param {Number} toolIndex - The index of the particular freehand tool.
-  * @param {Object} handleIndexArray - An array of indicies corresponding to the line segment.
-  * @return {Boolean} True if the mouse point projects onto the line segment
+  * @param {number} toolIndex The index of the particular freehand tool.
+  * @param {object} handleIndexArray An array of indicies corresponding to the line segment.
+  * @return {boolean} True if the mouse point projects onto the line segment
   */
   _pointProjectsToLineSegment (toolIndex, handleIndexArray) {
     const eventData = this._eventData;
@@ -216,11 +225,12 @@ export class FreehandLineFinder {
 
   /**
   * Returns the canvas positions from the handle's pixel positions.
+  * @static @public @method
   *
-  * @param {FreehandHandleData} handle1 - The first handle.
-  * @param {FreehandHandleData} handle2 - The second handle.
-  * @param {Object} element - The element on which the handles reside.
-  * @return {Object} An array contsining the handle positions in canvas coordinates.
+  * @param {FreehandHandleData} handle1 The first handle.
+  * @param {FreehandHandleData} handle2 The second handle.
+  * @param {object} element The element on which the handles reside.
+  * @returns {object} An array contsining the handle positions in canvas coordinates.
   */
   static getCanvasPointsFromHandles (handle1, handle2, element) {
     const p = [];
@@ -239,9 +249,10 @@ export class FreehandLineFinder {
 
   /**
   * Converts a line segment to a vector.
+  * @static @public @method
   *
-  * @param {Object} p - An array of two points respresenting the line segment.
-  * @return {Object} An array containing the x and y components of the vector, as well as a magnitude property.
+  * @param {object} p An array of two points respresenting the line segment.
+  * @return {object} An array containing the x and y components of the vector, as well as a magnitude property.
   */
   static getLineAsVector (p) {
     const r = [
@@ -256,9 +267,10 @@ export class FreehandLineFinder {
 
   /**
   * Constructs a vector from the direction and magnitude of the line from the the line origin to the mouse cursor.
+  * @private @method
   *
-  * @param {Object} p - An array of two points respresenting the line segment.
-  * @return {Object} An array containing the x and y components of the vector.
+  * @param {object} p An array of two points respresenting the line segment.
+  * @return {object} An array containing the x and y components of the vector.
   */
   _getLineOriginToMouseAsVector (p) {
     const m = [
@@ -271,10 +283,11 @@ export class FreehandLineFinder {
 
   /**
   * Calculates the perpendicular distance of the mouse cursor from a line segment.
+  * @private @method
   *
-  * @param {FreehandHandleData} handle1 - The first handle.
-  * @param {FreehandHandleData} handle2 - The first handle.
-  * @return {Number} The perpendicular distance of the mouse cursor from the line segment.
+  * @param {FreehandHandleData} handle1 The first handle.
+  * @param {FreehandHandleData} handle2 The first handle.
+  * @return {number} The perpendicular distance of the mouse cursor from the line segment.
   */
   _distanceOfPointfromLine (handle1, handle2) {
     const eventData = this._eventData;
@@ -295,10 +308,11 @@ export class FreehandLineFinder {
 
   /**
   * Gets the next handl index from a cyclical array of points.
+  * @static @public @method
   *
-  * @param {Number} currentIndex - The current index.
-  * @param {Number} length - The number of handles in the polygon.
-  * @return {Number} The index of the next handle.
+  * @param {number} currentIndex The current index.
+  * @param {number} length The number of handles in the polygon.
+  * @return {number} The index of the next handle.
   */
   static getNextHandleIndex (currentIndex, length) {
     let nextIndex;
