+++ src/util/freehand/calculateFreehandStatistics.js
@@ -1,15 +1,16 @@
 import pointInFreehand from './pointInFreehand.js';
 
 /**
-* Calculates the statistics of all the points within the freehand object.
-*
-* @param {Object} sp - An array of the pixel data.
-* @param {Object} boundingBox - Rectangular box enclosing the polygon.
-* @param {Object} dataHandles - Data object associated with the tool.
-* @return {Object} statisticsObj - Object containing the derived statistics.
-*/
+ * Calculates the statistics of all the pixels within the freehand object.
+ * @export @public @method
+ * @name calculateFreehandStatistics
+ *
+ * @param {object} sp An array of the pixel data.
+ * @param {object} boundingBox Rectangular box enclosing the polygon.
+ * @param {object} dataHandles Data object associated with the tool.
+ * @returns {object} Object containing the derived statistics.
+ */
 export default function (sp, boundingBox, dataHandles) {
-
   const statisticsObj = {
     count: 0,
     mean: 0.0,
@@ -32,13 +33,15 @@ export default function (sp, boundingBox, dataHandles) {
 }
 
 /**
-* Calculates the sum, squared sum and count of all pixels within the polygon.
-*
-* @param {Object} sp - An array of the pixel data.
-* @param {Object} boundingBox - Rectangular box enclosing the polygon.
-* @param {Object} dataHandles - Data object associated with the tool.
-* @return {Object} sum - Object containing the sum, squared sum and pixel count.
-*/
+ * Calculates the sum, squared sum and count of all pixels within the polygon.
+ * @private @method
+ * @name getSum
+ *
+ * @param {object} sp An array of the pixel data.
+ * @param {object} boundingBox Rectangular box enclosing the polygon.
+ * @param {object} dataHandles Data object associated with the tool.
+ * @return {object} Object containing the sum, squared sum and pixel count.
+ */
 function getSum (sp, boundingBox, dataHandles) {
   const sum = {
     value: 0,
@@ -63,14 +66,16 @@ function getSum (sp, boundingBox, dataHandles) {
 }
 
 /**
-* Adds the pixel to the workingSum if it is within the polygon.
-*
-* @param {Object} dataHandles - Data object associated with the tool.
-* @param {Object} point - The pixel coordinates.
-* @param {Object} workingSum - The working sum, squared sum and pixel count.
-* @param {Object} pixelValue - The pixel value.
-* @modifies {workingSum}
-*/
+ * Adds the pixel to the workingSum if it is within the polygon.
+ * @private @method
+ * @name sumPointIfInFreehand
+ *
+ * @param {object} dataHandles Data object associated with the tool.
+ * @param {object} point The pixel coordinates.
+ * @param {object} workingSum The working sum, squared sum and pixel count.
+ * @param {object} pixelValue The pixel value.
+ * @modifies {workingSum}
+ */
 function sumPointIfInFreehand (dataHandles, point, workingSum, pixelValue) {
   if (pointInFreehand(dataHandles, point)) {
     workingSum.value += pixelValue;
