+++ src/util/freehand/freehandIntersect.js
@@ -4,12 +4,14 @@
 */
 
 /**
-* Determines whether a new handle causes an intersection of the lines of the polygon.
-*
-* @param {Object} candidateHandle - The new handle to check.
-* @param {Object} dataHandles - data object associated with the tool.
-* @return {Boolean} - Whether the new line intersects with any other lines of the polygon.
-*/
+ * Determines whether a new handle causes an intersection of the lines of the polygon.
+ * @export @public @method
+ * @name newHandle
+ *
+ * @param {object} candidateHandle The new handle to check.
+ * @param {object} dataHandles data object associated with the tool.
+ * @returns {boolean} - Whether the new line intersects with any other lines of the polygon.
+ */
 function newHandle (candidateHandle, dataHandles) {
   // Check if the proposed line will intersect any existent line
   const lastHandleId = dataHandles.length - 1;
@@ -20,11 +22,13 @@ function newHandle (candidateHandle, dataHandles) {
 }
 
 /**
-* Checks if the last line of a polygon will intersect the other lines of the polgyon.
-*
-* @param {Object} dataHandles - data object associated with the tool.
-* @return {Boolean} - Whether the last line intersects with any other lines of the polygon.
-*/
+ * Checks if the last line of a polygon will intersect the other lines of the polgyon.
+ * @export @public @method
+ * @name end
+ *
+ * @param {object} dataHandles data object associated with the tool.
+ * @returns {boolean} Whether the last line intersects with any other lines of the polygon.
+ */
 function end (dataHandles) {
   const lastHandleId = dataHandles.length - 1;
   const lastHandle = getCoords(dataHandles[lastHandleId]);
@@ -34,12 +38,14 @@ function end (dataHandles) {
 }
 
 /**
-* Checks whether the modification of a handle's position causes intersection of the lines of the polygon
-*
-* @param {Object} dataHandles - data object associated with the tool.
-* @param {Number} modifiedHandleId - The id of the handle being modified.
-* @return {Boolean} - Whether the modfication causes any intersections.
-*/
+ * Checks whether the modification of a handle's position causes intersection of the lines of the polygon.
+ * @export @public @method
+ * @name modify
+ *
+ * @param {object} dataHandles Data object associated with the tool.
+ * @param {number} modifiedHandleId The id of the handle being modified.
+ * @return {boolean} Whether the modfication causes any intersections.
+ */
 function modify (dataHandles, modifiedHandleId) {
   // Check if the modifiedHandle's previous and next lines will intersect any other line in the polygon
   const modifiedHandle = getCoords(dataHandles[modifiedHandleId]);
@@ -70,14 +76,16 @@ function modify (dataHandles, modifiedHandleId) {
 }
 
 /**
-* Checks whether the line (p1,q1) intersects any of the other lines in the polygon.
-*
-* @param {Object} dataHandles - data object associated with the tool.
-* @param {Object} p1 - Coordinates of the start of the line.
-* @param {Object} q1 - Coordinates of the end of the line.
-* @param {Object} ignoredHandleIds - Ids of handles to ignore (i.e. lines that share a vertex with the line being tested).
-* @return {Boolean} - Whether the line intersects any of the other lines in the polygon.
-*/
+ * Checks whether the line (p1,q1) intersects any of the other lines in the polygon.
+ * @private @method
+ * @name doesIntersectOtherLines
+ *
+ * @param {object} dataHandles Data object associated with the tool.
+ * @param {object} p1 Coordinates of the start of the line.
+ * @param {object} q1 Coordinates of the end of the line.
+ * @param {object} ignoredHandleIds Ids of handles to ignore (i.e. lines that share a vertex with the line being tested).
+ * @return {boolean} Whether the line intersects any of the other lines in the polygon.
+ */
 function doesIntersectOtherLines (dataHandles, p1, q1, ignoredHandleIds) {
   let j = dataHandles.length - 1;
 
@@ -103,14 +111,16 @@ function doesIntersectOtherLines (dataHandles, p1, q1, ignoredHandleIds) {
 }
 
 /**
-* Checks whether the line (p1,q1) intersects the line (p2,q2) via an orientation algorithm.
-*
-* @param {Object} p1 - Coordinates of the start of the line 1.
-* @param {Object} q1 - Coordinates of the end of the line 1.
-* @param {Object} p1 - Coordinates of the start of the line 2.
-* @param {Object} q1 - Coordinates of the end of the line 2.
-* @return {Boolean} - Whether lines (p1,q1) and (p2,q2) intersect.
-*/
+ * Checks whether the line (p1,q1) intersects the line (p2,q2) via an orientation algorithm.
+ * @private @method
+ * @name doesIntersect
+ *
+ * @param {object} p1 Coordinates of the start of the line 1.
+ * @param {object} q1 Coordinates of the end of the line 1.
+ * @param {object} p1 Coordinates of the start of the line 2.
+ * @param {object} q1 Coordinates of the end of the line 2.
+ * @returns {boolean} Whether lines (p1,q1) and (p2,q2) intersect.
+ */
 function doesIntersect (p1, q1, p2, q2) {
   let result = false;
 
@@ -140,12 +150,14 @@ function doesIntersect (p1, q1, p2, q2) {
   return result;
 }
 
-/*
-* Returns an object with two properties, x and y, from a heavier FreehandHandleData object.
-*
-* @param {Object} dataHandle - data object associated with a single handle in the freehand tool.
-* @return {Object} - An object containing position propeties x and y.
-*/
+/**
+ * Returns an object with two properties, x and y, from a heavier FreehandHandleData object.
+ * @private @method
+ * @name getCoords
+ *
+ * @param {object} dataHandle Data object associated with a single handle in the freehand tool.
+ * @returns {object} An object containing position propeties x and y.
+ */
 function getCoords (dataHandle) {
   return {
     x: dataHandle.x,
@@ -154,13 +166,15 @@ function getCoords (dataHandle) {
 }
 
 /**
-* Checks the orientation of 3 points.
-*
-* @param {Object} p - First point.
-* @param {Object} q - Second point.
-* @param {Object} r - Third point.
-* @return {Number} - 0: Colinear, 1: Clockwise, 2: Anticlockwise
-*/
+ * Checks the orientation of 3 points.
+ * @private @method
+ * @name orientation
+ *
+ * @param {object} p First point.
+ * @param {object} q Second point.
+ * @param {object} r Third point.
+ * @returns {number} - 0: Colinear, 1: Clockwise, 2: Anticlockwise
+ */
 function orientation (p, q, r) {
   const orientationValue = (q.y - p.y) * (r.x - q.x) - (q.x - p.x) * (r.y - q.y);
 
@@ -172,13 +186,15 @@ function orientation (p, q, r) {
 }
 
 /**
-* Checks if point q lines on the segment (p,r).
-*
-* @param {Object} p - Point p.
-* @param {Object} q - Point q.
-* @param {Object} r - Point r.
-* @return {Boolean} - If q lies on line segment (p,r).
-*/
+ * Checks if point q lines on the segment (p,r).
+ * @private @method
+ * @name onSegment
+ *
+ * @param {object} p Point p.
+ * @param {object} q Point q.
+ * @param {object} r Point r.
+ * @returns {boolean} - If q lies on line segment (p,r).
+ */
 function onSegment (p, q, r) {
 
   if (q.x <= Math.max(p.x, r.x) && q.x >= Math.min(p.x, r.x) &&
