+++ src/util/freehand/index.js
@@ -1,0 +1,19 @@
+import calculateFreehandStatistics from './calculateFreehandStatistics.js';
+import ClickedLineData from './ClickedLineData.js';
+import freehandArea from './freehandArea.js';
+import FreehandHandleData from './FreehandHandleData.js';
+import freehandIntersect from './freehandIntersect.js';
+import FreehandLineFinder from './FreehandLineFinder.js';
+import insertOrDelete from './insertOrDelete.js';
+import pointInFreehand from './pointInFreehand.js';
+
+export default {
+  calculateFreehandStatistics,
+  ClickedLineData,
+  freehandArea,
+  FreehandHandleData,
+  freehandIntersect,
+  FreehandLineFinder,
+  insertOrDelete,
+  pointInFreehand
+};
